+++ haskell-src/Concordium/Types/Accounts.hs
@@ -3,6 +3,7 @@
 {-# LANGUAGE DeriveFunctor #-}
 {-# LANGUAGE DerivingVia #-}
 {-# LANGUAGE GADTs #-}
+{-# LANGUAGE InstanceSigs #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE ScopedTypeVariables #-}
@@ -66,10 +67,6 @@ module Concordium.Types.Accounts (
     delegationStakeEarnings,
     delegationTarget,
     delegationPendingChange,
-    CooldownTime (..),
-    CooldownTimeCode (..),
-    encodeCooldownTime,
-    decodeCooldownTime,
     AccountStake (..),
     AccountStakeHash (..),
     getAccountStakeHash,
@@ -85,14 +82,11 @@ module Concordium.Types.Accounts (
 
 import Data.Aeson
 import Data.Aeson.Types (Parser)
-import qualified Data.Bits as Bits
 import Data.Bool.Singletons
 import qualified Data.Map as Map
-import Data.Maybe
 import Data.Serialize
 import Data.Singletons
 import Data.Time
-import Data.Word
 import Lens.Micro.Platform (Lens', lens, makeClassy, makeLenses, (^.))
 
 import Concordium.Common.Version
@@ -276,6 +270,7 @@ deriving instance Ord (PendingChangeEffective av)
 deriving instance Show (PendingChangeEffective av)
 
 instance (IsAccountVersion av) => Serialize (PendingChangeEffective av) where
+    put :: (IsAccountVersion av) => Putter (PendingChangeEffective av)
     put (PendingChangeEffectiveV0 epoch) = put epoch
     put (PendingChangeEffectiveV1 timestamp) = put timestamp
     get = case delegationSupport @av of
@@ -394,147 +389,6 @@ instance forall av. (IsAccountVersion av, AVSupportsDelegation av) => Serialize 
         _delegationTarget <- get
         _delegationPendingChange <- get
         return AccountDelegationV1{..}
-
--- | Records
-data CooldownTime
-    = -- | In active cooldown. Timestamp must be less than 2^63
-      CooldownTimestamp !Timestamp
-    | -- | Will enter cooldown at next epoch, with the specified cooldown expiry timestamp.
-      --  Timestamp must be less than 2^62 - 1
-      PreCooldownTS !Timestamp
-    | -- | Will enter cooldown at next epoch, with the cooldown determined by the cooldown duration.
-      PreCooldown
-    | -- | Will enter cooldown at payday after next epoch, with the specified cooldown expiry
-      --  timestamp. Timestamp must be less than 2^62 - 1
-      PrePreCooldownTS !Timestamp
-    | -- | Will enter cooldown at payday after next epoch, with the cooldown determined by the
-      -- cooldown duration.
-      PrePreCooldown
-    deriving (Eq, Ord, Show)
-
--- | This type encodes a `CooldownTime` as a 64-bit integer. The encoding is designed to preserve
---  the ordering. Note that not all timestamps are representable, but it should be over a hundred
---  million years before that becomes relevant.  The table below shows the coding:
---
---   +-----------------------+-----------------------------------------+
---   | @theCooldownTimeCode@ | `CooldownTime`                          |
---   +=======================+=========================================+
---   | @0x0000000000000000@  | @CooldownTimestamp 0x0000000000000000@  |
---   +-----------------------+-----------------------------------------+
---   | ...                   | ...                                     |
---   +-----------------------+-----------------------------------------+
---   | @0x7fffffffffffffff@  | @CooldownTimestamp 0x7fffffffffffffff@  |
---   +-----------------------+-----------------------------------------+
---   | @0x8000000000000000@  | @PreCooldownTS 0x0000000000000000@      |
---   +-----------------------+-----------------------------------------+
---   | ...                   | ...                                     |
---   +-----------------------+-----------------------------------------+
---   | @0xbffffffffffffffe@  | @PreCooldownTS 0x3ffffffffffffffe@      |
---   +-----------------------+-----------------------------------------+
---   | @0xbfffffffffffffff@  | @PreCooldown@                           |
---   +-----------------------+-----------------------------------------+
---   | @0xc000000000000000@  | @PrePreCooldownTS 0x0000000000000000@   |
---   +-----------------------+-----------------------------------------+
---   | ...                   | ...                                     |
---   +-----------------------+-----------------------------------------+
---   | @0xfffffffffffffffe@  | @PrePreCooldownTS 0x3ffffffffffffffe@   |
---   +-----------------------+-----------------------------------------+
---   | @0xffffffffffffffff@  | @PrePreCooldown@                        |
---   +-----------------------+-----------------------------------------+
-newtype CooldownTimeCode = CooldownTimeCode {theCooldownTimeCode :: Word64}
-    deriving (Eq, Ord, Show)
-
-maxCooldownTimestampCode :: CooldownTimeCode
-maxCooldownTimestampCode = CooldownTimeCode 0x7fffffffffffffff
-
-encodeCooldownTime :: CooldownTime -> CooldownTimeCode
-encodeCooldownTime (CooldownTimestamp ts)
-    | ts < 0x8000000000000000 = CooldownTimeCode $ tsMillis ts
-encodeCooldownTime (PreCooldownTS ts)
-    | ts < 0x3fffffffffffffff = CooldownTimeCode $ tsMillis ts Bits..|. 0x8000000000000000
-encodeCooldownTime PreCooldown = CooldownTimeCode 0xbfffffffffffffff
-encodeCooldownTime (PrePreCooldownTS ts)
-    | ts < 0x3fffffffffffffff = CooldownTimeCode $ tsMillis ts Bits..|. 0xc000000000000000
-encodeCooldownTime PrePreCooldown = CooldownTimeCode 0xffffffffffffffff
-encodeCooldownTime _ = error "CooldownTime is not representable"
-
-decodeCooldownTime :: CooldownTimeCode -> CooldownTime
-decodeCooldownTime (CooldownTimeCode code) =
-    if Bits.testBit code 63
-        then
-            if Bits.testBit code 62
-                then
-                    if code == 0xffffffffffffffff
-                        then PrePreCooldown
-                        else PrePreCooldownTS . Timestamp $! code Bits..&. 0x3fffffffffffffff
-                else
-                    if code == 0xbfffffffffffffff
-                        then PreCooldown
-                        else PreCooldownTS . Timestamp $! code Bits..&. 0x3fffffffffffffff
-        else CooldownTimestamp (Timestamp code)
-
-instance Serialize CooldownTime where
-    put = putWord64be . theCooldownTimeCode . encodeCooldownTime
-    get = decodeCooldownTime . CooldownTimeCode <$> getWord64be
-
-data CooldownQueue (av :: AccountVersion) where
-    EmptyCooldownQueue :: CooldownQueue av
-    CooldownQueue ::
-        (SupportsFlexibleCooldown av ~ 'True) =>
-        -- | Entries in the map must be non-zero amounts, and the map must be non-empty.
-        Map.Map CooldownTimeCode Amount ->
-        CooldownQueue av
-
--- | Process all cooldowns that expire at or before the given timestamp.
---  If there are no such cooldowns, then 'Nothing' is returned.
---  Otherwise, the total amount exiting cooldown and the remaining queue are returned.
-processCooldowns :: Timestamp -> CooldownQueue av -> Maybe (Amount, CooldownQueue av)
-processCooldowns _ EmptyCooldownQueue = Nothing
-processCooldowns ts (CooldownQueue queue)
-    | freeAmount == 0 = Nothing
-    | otherwise = Just (freeAmount, remainder)
-  where
-    freeAmount = sum free + sum bonus
-    (free, bonus, keep) = Map.splitLookup (encodeCooldownTime (CooldownTimestamp ts)) queue
-    remainder
-        | null keep = EmptyCooldownQueue
-        | otherwise = CooldownQueue keep
-
-processPreCooldown :: Timestamp -> CooldownQueue av -> Maybe (CooldownQueue av)
-processPreCooldown _ EmptyCooldownQueue = Nothing
-processPreCooldown ts (CooldownQueue queue)
-    | null precooldowns = Nothing
-    | otherwise = Just . CooldownQueue $ Map.unionsWith (+) [cooldowns, newCooldowns, preprecooldowns]
-  where
-    (cooldowns, rest) = Map.spanAntitone (<= maxCooldownTimestampCode) queue
-    (precooldowns, preprecooldowns) = Map.spanAntitone (<= encodeCooldownTime PreCooldown) rest
-    newCooldowns = Map.mapKeysWith (+) f precooldowns
-    f c@(CooldownTimeCode code)
-        | c == encodeCooldownTime PreCooldown = CooldownTimeCode $ tsMillis ts
-        | otherwise = CooldownTimeCode (Bits.clearBit code 63)
-
-isCooldownQueueEmpty :: CooldownQueue av -> Bool
-isCooldownQueueEmpty EmptyCooldownQueue = True
-isCooldownQueueEmpty _ = False
-
-nextCooldownTime :: CooldownQueue av -> Maybe Timestamp
-nextCooldownTime EmptyCooldownQueue = Nothing
-nextCooldownTime (CooldownQueue queue) = case decodeCooldownTime minEntry of
-    CooldownTimestamp ts -> Just ts
-    _ -> Nothing
-  where
-    -- This is safe because 'CooldownQueue' requires @queue@ to be non-empty.
-    (minEntry, _) = Map.findMin queue
-
-hasPreCooldown :: CooldownQueue av -> Bool
-hasPreCooldown EmptyCooldownQueue = False
-hasPreCooldown (CooldownQueue queue) = case Map.lookupGT maxCooldownTimestampCode queue of
-    Just (x, _) -> x <= encodeCooldownTime PreCooldown
-    Nothing -> False
-
-hasPrePreCooldown :: CooldownQueue av -> Bool
-hasPrePreCooldown EmptyCooldownQueue = False
-hasPrePreCooldown (CooldownQueue queue) = isJust $ Map.lookupGT (encodeCooldownTime PreCooldown) queue
 
 -- | Whether an account stakes as a baker, delegates to a baker, or neither.
 data AccountStake (av :: AccountVersion) where
